tests for section values
